VFS Fixes, Maven, Unit Tests
