+++ .travis.yml
@@ -1,0 +1,4 @@
+language: java
+cache:
+  directories:
+  - $HOME/.m2
