--- build.xml
@@ -1,128 +1,0 @@
-<project name="MaxMindGeoIPLookup" default="default" basedir=".">
-
-	  <description>
-	      MaxMindGeoIPLookup plugin build file
-	    </description>
-
-	  <tstamp prefix="start"/>
-
-	  <property file="override.properties"/>
-	  <property name="libswt" value="libswt"/>
-	  <property name="libext" value="libext"/>
-
-	  <!-- Plugin Name -->
-	  <property name= "plugin-name" value="MaxMindGeoIPLookup"/> 
-
-	  <!-- system properties to use -->
-	  <property name= "cr" value="${line.separator}"/> 
-		
-	  <!-- set global properties for this build -->
-	
-	  <property name="src"        			location="src"/> 
-	  <property name="classes"        		location="classes"/> 
-	  <property name="lib"        			location="lib"/>
-	  <property name="samples"        		location="samples"/>
-	  <property name="distrib"        		location="distrib"/>
-	
-	  <target name="init" description="create timestamp and directories">
-    	<echo>Init...</echo>
-	  	
-	    <tstamp/>
-	    <mkdir dir="${classes}"/>
-	    <mkdir dir="${distrib}"/>
-	    <mkdir dir="${deploy}"/>
-	  </target>
-
-	  <!--
-	       ==========================
-	            Compile the code
-	       ==========================
-	  -->
-
-	  <target name="compile" depends="init" description="compile the source " >
-	  	<echo>Compiling MaxMind GeoIP Lookup...</echo>
-	    <javac debug="true" srcdir="${src}" 
-	    	   destdir="${classes}"
-                   includeantruntime="false"
-                   >
-
-	    	<classpath id="cpath">
-				<fileset dir="${lib}" 		        includes="*.jar"/>
-				<fileset dir="${libext}" 		includes="*.jar"/>
-				<fileset dir="${libswt}" 		includes="*.jar"/>
-		  		</classpath>
-	    </javac>
-	  </target>
-	
-	  <!-- ==========================
-	          Copy additional files
-	       ==========================
-	  -->
-
-	  <target name="copy" depends="compile" description="copy dependend libraries" >
-	  	<echo>Copying required libraries...</echo>
-	    <copy todir="${distrib}/lib"> <fileset dir="${lib}/" includes="**/*"/> </copy>
-	    <copy todir="${distrib}/samples"> <fileset dir="${samples}/" includes="**/*"/> </copy>
-	    <copy todir="${distrib}/"> <fileset dir="." includes="README.txt"/> </copy>
-	    <copy todir="${classes}"> <fileset dir="${src}" includes="**/*.png **/*.xml **/*.xsl **/*properties"/> </copy>
-	  </target>
-			
-	  <!--
-	       ==========================
-	                library
-	       ==========================
-	  -->
-		
-	  <target name="jar" depends="compile, copy" description="generate the MaxMind GeoIP Lookup library jar">
-	  	<echo>Generating the MaxMindGeoIPLookup library jar ...</echo>
-	  	<jar 
-	    	jarfile="${distrib}/maxmindgeoiplookup.jar" 
-	    	basedir="${classes}" 
-	    	includes="**/*"/>	 
-	  </target>
-
-	  <!--
-	       ==========================
-	                Plugin ZIP
-	       ==========================
-	  -->
-		
-	  <target name="zip" depends="jar" description="generate the MaxMind GeoIP Lookup plugin zip">
-	  	<echo>Generating the MaxMindGeoIPLookup plugin zip ...</echo>
-	  	<zip 
-	    	zipfile="maxmindlookup.zip" 
-	    	basedir="${distrib}" 
-	    	includes="**/*"/>
-	  </target>
-
-	  <!-- ======================================
-	          Copy data to a deployment folder
-	       ======================================
-	  -->
-
-	  <target name="deploy" depends="zip" description="deploy as a Kettle plugin" >
-	  	<echo>Deploying as a Kettle plugin...</echo>
-	    <copy todir="${deploy}"> <fileset dir="${distrib}" includes="**/*"/> </copy>
-	  </target>
-			
-	<!--
-	       ==========================
-	          Default target: all
-	       ==========================
-	  -->
-	<target name="default" depends="deploy" description="default = build all"/>
-
-	<!--
-	       ==========================
-	              C L E A N
-	       ==========================
-	  -->
-
-	  <target name="clean"  description="clean up generated files" >
-	    <delete dir="${classes}"/>
-	    <delete dir="${distrib}"/>
-	    <delete file="maxmindlookup.zip"/>
-	    <delete dir="${deploy}"/>
-	  </target>
-	
-</project>
