--- src/com/maxmind/geoip/messages/messages_en_US.properties
@@ -1,12 +1,0 @@
-MaxMindGeoIPLookupDialog.Shell.Title=MaxMind GeoIP lookup
-MaxMindGeoIPLookupDialog.StepName.Label=Step Name
-MaxMindGeoIPLookupDialog.IPAddressFieldName.Label=IP Address Field Name
-MaxMindGeoIPLookupDialog.DBLocation.Label=DB Location
-MaxMindGeoIPLookupDialog.DBLocation.Tooltip=Path to the MaxMind GeoIP binary database file
-MaxMindGeoIPLookupDialog.Fields.Label=Fields
-MaxMindGeoIPLookupDialog.ColumnInfo.NewField=Name
-MaxMindGeoIPLookupDialog.ColumnInfo.LookupType=Lookup Type
-MaxMindGeoIPLookupDialog.ColumnInfo.IfNull=Default
-MaxMindGeoIPLookupMeta.TypeTooltipDesc.MaxMindGeoIPLookup=MaxMind GeoIP lookup
-MaxMindGeoIPLookupMeta.TypeLongDesc.MaxMindGeoIPLookup=GeoIP lookup
-MaxMindGeoIPLookupMeta.Category=Lookup
